Time: 6 ms (82.62%), Space: 44.6 MB (68.25%) - LeetHub

+++ 0462-minimum-moves-to-equal-array-elements-ii/0462-minimum-moves-to-equal-array-elements-ii.java
@@ -4,7 +4,7 @@ class Solution {
         int mid = nums[nums.length / 2];
         int ans = 0;
         for (int i : nums) {
-            ans += (int) Math.abs(i - mid);
+            ans += Math.abs(i - mid);
         }
         return ans;
     }
